Module Irmin_fs

Disk persistence.

val config : ?config:Irmin.config -> string -> Irmin.config

config ?config root is the configuration config augmented with the key Irmin.Config.root set to root. If not specified, config is Irmin.Config.empty.

module type IO = sig ... end

module Append_only (IO : IO) : Irmin.APPEND_ONLY_STORE_MAKER

module Atomic_write (IO : IO) : Irmin.ATOMIC_WRITE_STORE_MAKER

module Make (IO : IO) : Irmin.S_MAKER

module KV (IO : IO) : Irmin.KV_MAKER

Advanced configuration

module type Config = sig ... end

Same as Config but gives more control on the file hierarchy.

module Append_only_ext (IO : IO) (C : Config) : Irmin.APPEND_ONLY_STORE_MAKER

module Atomic_write_ext (IO : IO) (C : Config) : Irmin.ATOMIC_WRITE_STORE_MAKER

module Make_ext (IO : IO) (Obj : Config) (Ref : Config) : Irmin.S_MAKER

In-memory IO mocks

module IO_mem : sig ... end