Module Irmin_fs.Append_only_ext

Parameters

module IO : IO

module C : Config

module K : Irmin.Type.S

module V : Irmin.Type.S

Signature

include Irmin.APPEND_ONLY_STORE with type key = K.t and type value = V.t

type 'a t

The type for append-only backend stores. The 'a phantom type carries information about the store mutability.

type key = K.t

The type for keys.

type value = V.t

The type for raw values.

val mem : [> `Read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> `Read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> `Write ] t -> key -> value -> unit Lwt.t

Write the contents of a value to the store.

val batch : [ `Read ] t -> ([ `Read | `Write ] t -> 'a Lwt.t) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depends on the backends.

val v : Irmin.config -> [ `Read ] t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated to t. Any operations run on a closed store will raise Closed.