Module Irmin_fs

Disk persistence.

module Conf : sig ... end

val config : string -> Irmin.config

config root is the a configuration with the key Irmin.Config.root set to root. *

module type IO = sig ... end

module Append_only (IO : IO) : Irmin.Append_only.Maker

module Atomic_write (IO : IO) : Irmin.Atomic_write.Maker

module Maker (IO : IO) : Irmin.Maker

module KV (IO : IO) : Irmin.KV_maker with type info = Irmin.Info.default

Advanced configuration

module type Config = sig ... end

Same as Config but gives more control on the file hierarchy.

module Append_only_ext (IO : IO) (C : Config) : Irmin.Append_only.Maker

module Atomic_write_ext (IO : IO) (C : Config) : Irmin.Atomic_write.Maker

module Maker_ext (IO : IO) (Obj : Config) (Ref : Config) : Irmin.Maker

In-memory IO mocks

module IO_mem : sig ... end